refactor: rename actions files
