+++ src/views/contacts/actions/tests/use-contact-preview-actions.test.tsx
@@ -6,7 +6,7 @@
 
 import { setupHook } from '../../../../carbonio-ui-commons/test/test-setup';
 import { buildContact } from '../../../../tests/model-builder';
-import { useContactPreviewActions } from '../contact-preview-actions';
+import { useContactPreviewActions } from '../use-contact-preview-actions';
 
 describe('Contact Preview Actions', () => {
 	it('should return [send, tag, edit, move, delete] actions in this order', () => {
